Tests/LibWeb/Text/expected/Cache/http-memory-cache.txt

Original response: "abcdefghijklmnabcdefghijklmnopqrstuvwxyz"
Range response: "klmnabcdef"
